pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py2femm"
version = "0.2.0"
description = "Python automation platform for FEMM finite element simulations"
readme = "README.md"
authors = [{ name = "Riccardo Tinivella", email = "tinix84@gmail.com" }]
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pandas>=1.5",
    "httpx>=0.24",
]

[project.optional-dependencies]
agent = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "httpx>=0.24",
]
all = ["py2femm[agent,dev]"]

[project.scripts]
py2femm = "py2femm.cli:main"

[tool.setuptools.packages.find]
include = ["py2femm*", "py2femm_agent*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
